+++ R/ms_update_new_observation.R
@@ -1,1 +1,0 @@
-ms_update_new_observation <- NULL